requirements.txt
PyJWT
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
marshmallow
marshmallow-sqlalchemy
python-consul